Update hat-tip in changelog
I don't want to claim any undue credit to tracking this error down to
the 0.1.8 release. Update the CHANGELOG to reflect that it was
@gisenberg who filed
https://github.com/labocho/thor-zsh_completion/issues/7

Honestly this is sort of a joke PR which I don't expect to be merged
(much less have a new version bumped out!), but I probably wouldn't
have looked into this far enough to have tracked down @labocho's
project and to have filed the issue above. I'd already discovered that
0.1.5 worked fine for me and was inclined to just move on with my day.

Real credit to @gisenberg, who cared enough to dig into the bug and
file it.